Moved data-related classes to separate data/ dir
