+++ kguitar/songview.cpp
@@ -8,8 +8,8 @@
 #include "tracklist.h"
 #include "trackpane.h"
 #include "trackdrag.h"
-#include "tabsong.h"
-#include "tabtrack.h"
+#include "data/tabsong.h"
+#include "data/tabtrack.h"
 #include "settrack.h"
 #include "settabfret.h"
 #include "settabdrum.h"
